index & show methods

+++ wisebits/routes/api.php
@@ -5,10 +5,10 @@ use Laravel\Lumen\Routing\Router;
 /** @var Router $router */
 $router->group(['prefix' => 'users'], function (Router $router) {
     $router->get('/', 'UserController@index');
-    $router->get('/{userId}', 'UserController@show');
+    $router->get('/{userId:[0-9]+}', 'UserController@show');
     $router->post('/', 'UserController@store');
-    $router->put('/', 'UserController@update');
-    $router->delete('/', 'UserController@delete');
+    $router->put('/{userId:[0-9]+}', 'UserController@update');
+    $router->delete('/{userId:[0-9]+}', 'UserController@delete');
 });
